Move MockLeapHybridSolver to testing.py
