+++ dwave/system/testing.py
@@ -17,11 +17,16 @@ import collections
 
 import dimod
 import dwave_networkx as dnx
+from tabu import TabuSampler
+from dwave.cloud.computation import Future as cloud_future
 
 try:
     from neal import SimulatedAnnealingSampler
 except ImportError:
     from dimod import SimulatedAnnealingSampler
+
+from concurrent.futures import Future
+import numpy as np
 
 try:
     # py3
@@ -29,7 +34,6 @@ try:
 except ImportError:
     # py2
     import mock
-
 
 C4 = dnx.chimera_graph(4, 4, 4)
 
@@ -89,3 +93,29 @@ class MockDWaveSampler(dimod.Sampler, dimod.Structured):
         return dimod.SampleSet.from_samples_bqm([{v: sample[v] for v in bqm}
                                                  for sample in response.samples()],
                                                 bqm)
+
+class MockLeapHybridSolver:
+
+    properties = {'supported_problem_types': ['bqm'],
+                  'minimum_time_limit': [[1, 1.0], [1024, 1.0],
+                                         [4096, 10.0], [10000, 40.0]],
+                  'parameters': {'time_limit': None},
+                  'category': 'hybrid',
+                  'quota_conversion_rate': 1}
+
+    def upload_bqm(self, bqm, **parameters):
+        bqm_adjarray = dimod.serialization.fileview.load(bqm)
+        future = Future()
+        future.set_result(bqm_adjarray)
+        return future
+
+    def sample_bqm(self, sapi_problem_id, time_limit):
+        #Workaround until TabuSampler supports C BQMs
+        bqm = dimod.BQM(sapi_problem_id.linear,
+                                    sapi_problem_id.quadratic,
+                                    sapi_problem_id.offset,
+                                    sapi_problem_id.vartype)
+        result = TabuSampler().sample(bqm, timeout=1000*int(time_limit))
+        future = cloud_future('fake_solver', None)
+        future._result = {'sampleset': result, 'problem_type': 'bqm'}
+        return future
