+++ tests/test_leaphybridsolver.py
@@ -14,14 +14,9 @@
 #
 # =============================================================================
 import unittest
-from concurrent.futures import Future
 import numpy as np
 
 import dimod
-from tabu import TabuSampler
-from dwave.system.samplers import LeapHybridSampler
-
-from dwave.cloud.computation import Future as cloud_future
 from dwave.cloud import Client
 
 try:
@@ -31,31 +26,8 @@ except ImportError:
     # py2
     import mock
 
-class MockLeapHybridSolver:
-
-    properties = {'supported_problem_types': ['bqm'],
-                  'minimum_time_limit': [[1, 1.0], [1024, 1.0],
-                                         [4096, 10.0], [10000, 40.0]],
-                  'parameters': {'time_limit': None},
-                  'category': 'hybrid',
-                  'quota_conversion_rate': 1}
-
-    def upload_bqm(self, bqm, **parameters):
-        bqm_adjarray = dimod.serialization.fileview.load(bqm)
-        future = Future()
-        future.set_result(bqm_adjarray)
-        return future
-
-    def sample_bqm(self, sapi_problem_id, time_limit):
-        #Workaround until TabuSampler supports C BQMs
-        bqm = dimod.BQM(sapi_problem_id.linear,
-                                    sapi_problem_id.quadratic,
-                                    sapi_problem_id.offset,
-                                    sapi_problem_id.vartype)
-        result = TabuSampler().sample(bqm, timeout=1000*int(time_limit))
-        future = cloud_future('fake_solver', None)
-        future._result = {'sampleset': result, 'problem_type': 'bqm'}
-        return future
+from dwave.system.samplers import LeapHybridSampler
+from dwave.system.testing import MockLeapHybridSolver
 
 # Called only for named solver
 class MockBadLeapHybridSolver:
